app/requirements.txt
flask~=3.1.1
protobuf~=5.29.4
PyYAML~=6.0.2
urllib3~=1.26.20
pydantic